[Fixes #19327703] Refactored numberhelper to create a currencynumberhelper (helpers for currencies)
