+++ db/migrate/20111005133024_move_sub_activities_to_implementer_splits.rb
@@ -40,7 +40,7 @@ class SubActivity < Activity
   end
 
   def budget=(amount)
-    if is_number?(amount)
+    if NumberHelper.is_number?(amount)
       write_attribute(:budget, amount.to_f.round_with_precision(2))
     else
       write_attribute(:budget, amount)
@@ -48,7 +48,7 @@ class SubActivity < Activity
   end
 
   def spend=(amount)
-    if is_number?(amount)
+    if NumberHelper.is_number?(amount)
       write_attribute(:spend, amount.to_f.round_with_precision(2))
     else
       write_attribute(:spend, amount)
