+++ lib/autocreate_helper.rb
@@ -1,6 +1,6 @@
 module AutocreateHelper
   def assign_or_create_organization(id_or_name)
-    if is_number?(id_or_name)
+    if NumberHelper.is_number?(id_or_name)
       new_id = id_or_name
     else
       organization = Organization.find_or_create_by_name(id_or_name)
